Resolve #12 - Lists events have index
When receiving an add/remove event from a list,
the event arguments may be cast to IHasListIndex in order
to get the zero-based index at which the item was added or
removed (or is to be added/removed).

A future change might alter the event API, but this was
done requiring a cast so as not to force a breaking change.